refactor: delete scheduling_helper init and restore back to origin
