+++ dags/tpu_observability/jobset_ttr_kill_process.py
@@ -39,7 +39,7 @@ from dags.tpu_observability.configs.common import (
     GCS_CONFIG_PATH,
     GCS_JOBSET_CONFIG_PATH,
 )
-from dags.common.scheduling_helper import SchedulingHelper, get_dag_timeout
+from dags.common.scheduling_helper.scheduling_helper import SchedulingHelper, get_dag_timeout
 
 
 DAG_ID = "jobset_ttr_kill_process"
