+++ dags/tpu_observability/node_pool_ttr_disk_size.py
@@ -25,7 +25,7 @@ from airflow.utils.trigger_rule import TriggerRule
 from dags import composer_env
 from dags.tpu_observability.configs.common import MachineConfigMap, GCS_CONFIG_PATH
 from dags.tpu_observability.utils import node_pool_util as node_pool
-from dags.common.scheduling_helper import SchedulingHelper, get_dag_timeout
+from dags.common.scheduling_helper.scheduling_helper import SchedulingHelper, get_dag_timeout
 
 
 DAG_ID = "node_pool_ttr_disk_size"
